Support active leadership takeover (#50)
* Once "yield leadership" is requested, the current leader chooses the
next leader. It pauses incoming writes, and waits until the next leader
candidate fully catches up the current latest log. After that, the
current leader sends a custom notification to that candidate.

* Once the next leader candidate receives the notification, it
immediately initiates a special vote which will ignore the current
priority, so that it can be the next leader quickly.