Fix to bug regarding stepping down newly added node
* Once a new node is added to existing cluster, it should clear the
`catching_up_` flag only after it is successfully added to the
cluster member list (i.e., config commit). Otherwise, there might be
any prior membership change configs that is already outdated so that
does not contain the new node. Then new node will treat it as an
abandon from the cluster so that step down itself.