Support global workers for commit and append (#136)
* Each Raft instance creates at least two threads: one for commit
and the other one for append. If there is a use case that a process
has many Raft instances, it will produce lots of threads that end
up with inefficient resource utilization.

* To address this issue, we introduce a global thread pool; there are
shared worker threads for commit and append, and a global manager
coordinates requests from multiple Raft instances and assigns them
to proper worker threads.