Joining member should ignore priority
* When a new member is added, its election timer is disabled until it
fully catches up with the leader. Suppose somehow the leader has a
problem so that a leader election is initiated. In that case, the
new member may refuse the vote request if the candidate's priority is
lower than its priority. Since the election timer is disabled, there
is no way to decrease the target priority of the new member;
consequently, the leader election will not succeed forever.

* To avoid such a situation, the new member in catch-up mode should
ignore priority for the vote.