Use find_package for OpenSSL detection
This refactors CMakeLists.txt to call find_package() instead of using
our own detection logic. The result is easier to read, and will benefit
from improvements to the battle-hardened FindOpenSSL module.

The following options' behavior was preserved:
DEPS_PATH, OPENSSL_INCLUDE_PATH, and OPENSSL_LIBRARY_PATH.

The change was implemented on OS X, so the developer was able to verify
that Homebrew's OpenSSL1.1 can be detected after the change.

This is continuing the work started in #358.