Clean up logic for finding OpenSSL
This divides up the logic into two distinct cases captured by macros.
The first represents the case where the library and include paths are
provided, and the second represents the case where we must search for
the library. The paths will be printed out appropriately in both cases,
and in the search case, a notice will be printed if OpenSSL is not
found.