Use find_package for OpenSSL detection (#451)
* Use find_package for OpenSSL detection

This refactors CMakeLists.txt to call find_package() instead of using
our own detection logic. The result is easier to read, and will benefit
from improvements to the battle-hardened FindOpenSSL module.

The following options' behavior was preserved:
DEPS_PATH, OPENSSL_INCLUDE_PATH, and OPENSSL_LIBRARY_PATH.

The change was implemented on OS X, so the developer was able to verify
that Homebrew's OpenSSL1.1 can be detected after the change.

This is continuing the work started in #358.

* Mark OpenSSL_XXX vars as advanced and cache them

It is risky to have seemingly magic behavior; the variables should
be in the cache so that they can be inspected.

* Apply feedback from greensky

- Remove extraneous space in status string
- Use static OpenSSL libraries

* Clean up logic for finding OpenSSL

This divides up the logic into two distinct cases captured by macros.
The first represents the case where the library and include paths are
provided, and the second represents the case where we must search for
the library. The paths will be printed out appropriately in both cases,
and in the search case, a notice will be printed if OpenSSL is not
found.

---------

Co-authored-by: Jung-Sang Ahn <jungsang.ahn@gmail.com>