Fix public usage requirements on CMake targets
This adds the include directory to the public includes for both the
static and shared targets. It is necessary to do it for both the build
and install interfaces, because the build interface will be used if the
project is added as a submodule, but the install interface will be used
if the project was added using a config.

Fixes #460.