Support new joiner functionality
* A new setting, `use_new_joiner_type_`, has been introduced, which
defaults to false.

* When activated, any new members are immediately added to the cluster
configuration as a "new joiner." Similar to learners, new joiners do
not participate in the quorum; they cannot initiate or partake in voting.

* Once the log gap of a new joiner narrows to less than
`log_sync_stop_gap_`, the member transitions to a regular status and
becomes eligible to engage in the leader election process.

* The advantage of this feature is that it allows a newly elected leader
to continue the synchronization process with the new joiner without
needing to invoke the `add_srv()` API again.