Fix: The snapshot installation blindly applies config to the server. (#566)
* Fix: The snapshot installation blindly applies config to the server.

If the server was previously in the member set which was captured in the config of
this snapshot, then the peer will appear to have "caught_up" in ::reconfigure. I do
not believe there is value in rolling back the config just to reapply it later once
log_sync occurs. `::commit_conf` performs a similar check.

* Simplify

* [Update PR] Leave a log when snapshot config is skipped

---------

Co-authored-by: Jung-Sang Ahn <jungsang.ahn@gmail.com>